animation: Squelch 'unknown option' error if other animation plugins are pending load
When defining an option for additional animation plugins, an error
is emmited because the option is parsed when animation plugin is
loaded, before other animation plugins. Squelch the message if other
animations are in the plugin list to be loaded but aren't loaded yet.
The error should now appear only once, after all animation plugins
have been loaded and the option has still not been matched.