ezoom: Fix painting on the fullscreen output
This fixes support of e.g. XRandr-cloned outputs.  It should fix any
situation where outputs overlap, in which case rendering is done on the
fullscreen output instead of per-output.

This makes the plugin properly handle the fullscreen output ID, and use
the zoom area of the first output device in that case.

Based on a patch from Samuel Thibault.