colorfilter: Add an option to enable cumulative filter mode
This is useful to control which filter is enabled at startup, and
choose whether to have cumulative mode or a specific filter.

Moreover, cumulative mode is interesting only when a specific set of
filters was selected, as not all filters play well together.

This changes the default from enabled to disabled, but it seems
reasonable as the cumulative mode is less commonly useful, and we even
currently have a broken filter that prevents the cumulative mode from
working at all.