titleinfo: don't flag remote windows as root-owned
The _NET_WM_PID property set on remote windows is the PID of the process
on the remote machine, so we cannot use it locally to determine whether
the process is root-owned.  As we have no way of discovering the
ownership of the remote process, we should just refuse to flag remote
windows as root-owned.