Remove leftover initiate option
It seems that the "initiate" sort-of standard option which is not used anymore
in ezoom is still there and by some default somewhere, <enter>/<escape> trigger
the terminate (un-initiate) action. Nothing in config allows to modify this
and we probably don't want to re-introduce an initiate step. Moreover it just
resets the zoom to 1.0 which can already be done by "specific zoom" actions.
Remove that stuff. Fixes issue #44.