Add a photo-sensitivity filter
This adds 2 filters actually: one only adjusting the colors, and the
other combining the color adjustment with negative-lightness.

These filters are not based on much data but on the feedback from a
user suffering from visual strain.

The photosensitivity-1 filter is the base transformation, mostly useful
in combination with one of the negative filters (negative-lightness
would be the go-to recommendation).

photosensitivity-1-negative is the same filter but combined with
negative-lightness for a ready-to-go experience, which plays nicer with
dynamic filter change.