ezoom: Fix crash when plugging in a new output
We need to update the list of zoom areas we maintain per output to
actually match the available outputs.  Failure to do so means we might
try and reference an area for an output we don't know about, leading to
a crash.

While at it, guard against a NULL zoom area in a couple of places,
although now we properly update to output changes it should not happen
short of memory allocation failure.  There still are a couple of places
that are not guarded properly, but what to fallback to there is more
questionable, and again, the root problem leading to a missing area for
an existing output is fixed now anyway.