focuspoll: Ignore region-changed events when not in say-all mode
These events when in e.g. focus mode are problematic because they have
a tendency to shadow regular focus events and thus moving the focus to
unexpected locations.  We didn't expect region-changed to be emitted
outside say-all mode, but upstream Orca generalized that, which forces
us to filter the non-interesting ones out.